requirements.txt
pandas==2.2.0
pandas-ta==0.3.14b
websockets==12.0
deriv-api==0.1.7
PyYAML==6.0.1
aiohttp==3.9.3
aioconsole==0.7.0
